refactor: move Model(Location, Place) to Domain
