+++ app/src/main/java/campus/tech/kakao/map/data/repository/DefaultLocationRepository.kt
@@ -2,7 +2,7 @@ package campus.tech.kakao.map.data.repository
 
 import android.util.Log
 import androidx.datastore.core.DataStore
-import campus.tech.kakao.map.data.model.Location
+import campus.tech.kakao.map.domain.model.Location
 import campus.tech.kakao.map.di.LocationDataStore
 import dagger.hilt.android.scopes.ViewModelScoped
 import kotlinx.coroutines.flow.firstOrNull
