+++ app/src/main/java/campus/tech/kakao/map/di/DatabaseModule.kt
@@ -5,7 +5,7 @@ import androidx.datastore.core.DataStore
 import androidx.datastore.dataStore
 import androidx.room.Room
 import campus.tech.kakao.map.data.database.AppDatabase
-import campus.tech.kakao.map.data.model.Location
+import campus.tech.kakao.map.domain.model.Location
 import campus.tech.kakao.map.data.repository.LocationSerializer
 import dagger.Module
 import dagger.Provides
