+++ app/src/main/java/campus/tech/kakao/map/ui/map/LocationViewModel.kt
@@ -3,7 +3,7 @@ package campus.tech.kakao.map.ui.map
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import campus.tech.kakao.map.data.model.Location
+import campus.tech.kakao.map.domain.model.Location
 import campus.tech.kakao.map.di.IoDispatcher
 import campus.tech.kakao.map.domain.usecase.LoadLocationUseCase
 import campus.tech.kakao.map.domain.usecase.SaveLocationUseCase
