+++ app/src/main/java/campus/tech/kakao/map/ui/search/interfaces/OnPlaceItemClickListener.kt
@@ -1,6 +1,6 @@
 package campus.tech.kakao.map.ui.search.interfaces
 
-import campus.tech.kakao.map.data.model.Place
+import campus.tech.kakao.map.domain.model.Place
 
 /**
  * 검색 결과 아이템 클릭 이벤트를 처리하는 인터페이스.
